fix-feat: Overhaul current way of page transition
Initially, the problem that I was trying to solve was allowing data of
previous pages to persist. This was so I could ensure that the buttons
that were disabled would stay disabled. I could have simply made global
variables to ensure persistence of such states that I could just compare
to, but instead I overhauled the page transition system from using
pageName.destroy to pageName.pack_forget.

This brought additional problems such as menu items and receipt items
not refreshing properly, which was solved through a simple deletion and
re-initialization of the items or the frames containing such items